+++ solmate_mqtt.py
@@ -13,9 +13,8 @@ from paho.mqtt.packettypes import PacketTypes
 
 class solmate_mqtt():
 
-	def __init__(self, merged_config, api_available):
+	def __init__(self, api_available):
 		# initialize MQTT with parameters
-		self.merged_config=merged_config
 		self.api_available = api_available
 		self.connect_ok = None
 		self.signal_reason = 0
@@ -30,17 +29,17 @@ class solmate_mqtt():
 		self.has_ha_config = False
 		self.first_query_has_run = False
 
-		utils.logging('Initializing the MQTT class.', self.merged_config)
+		utils.logging('Initializing the MQTT class.')
 
 		# get the merged config and use mqtt relevant settings
-		self.mqtt_server = self.merged_config['mqtt_server']
-		self.mqtt_port = int(self.merged_config['mqtt_port'])
-		self.mqtt_username = self.merged_config['mqtt_username']
-		self.mqtt_password = self.merged_config['mqtt_password']
-		self.mqtt_client_id = self.merged_config['mqtt_client_id']
-		self.mqtt_topic = self.merged_config['mqtt_topic']
-		self.mqtt_prefix = self.merged_config['mqtt_prefix']
-		self.mqtt_ha = self.merged_config['mqtt_ha']
+		self.mqtt_server = utils.merged_config['mqtt_server']
+		self.mqtt_port = int(utils.merged_config['mqtt_port'])
+		self.mqtt_username = utils.merged_config['mqtt_username']
+		self.mqtt_password = utils.merged_config['mqtt_password']
+		self.mqtt_client_id = utils.merged_config['mqtt_client_id']
+		self.mqtt_topic = utils.merged_config['mqtt_topic']
+		self.mqtt_prefix = utils.merged_config['mqtt_prefix']
+		self.mqtt_ha = utils.merged_config['mqtt_ha']
 
 		# https://www.home-assistant.io/integrations/mqtt/#discovery-topic
 		# <discovery_prefix>/<component>/[<node_id>/]<object_id>/config
@@ -78,7 +77,7 @@ class solmate_mqtt():
 	def init_mqtt_client(self):
 		try:
 			# initialize the MQTT client. to see if it was successful, you must go to _on_connect
-			utils.logging('Initializing the MQTT client.', self.merged_config)
+			utils.logging('Initializing the MQTT client.')
 
 			# protocol versions available
 			# MQTTv31  = 3
@@ -124,7 +123,7 @@ class solmate_mqtt():
 			)
 			self.mqttclient.loop_start()
 
-			#utils.logging('MQTT connection failed: ' + str(err), self.merged_config)
+			#utils.logging('MQTT connection failed: ' + str(err))
 			#sys.exit()
 
 			# wait until on_connect returns a response
@@ -140,11 +139,11 @@ class solmate_mqtt():
 		except Exception as err:
 			# print any other error that has occured but only if it was not raised by us
 			if len(err.args) != 2:
-				utils.logging(str(err), self.merged_config)
+				utils.logging(str(err))
 			raise Exception('mqtt', 'timer_offline')
 
 		# update HA topics to initialise correctly
-		utils.logging('Update MQTT topics for Homeassistant.', self.merged_config)
+		utils.logging('Update MQTT topics for Homeassistant.')
 
 		# update the home assistant auto config info
 		# each item needs its own publish
@@ -205,7 +204,7 @@ class solmate_mqtt():
 		# the 'will_set' is not sent on graceful shutdown by design
 		# we need to wait until the message has been sent, else it will not appear in the broker
 		if self.connect_ok:
-			utils.logging('\rShutting down MQTT gracefully.', self.merged_config)
+			utils.logging('\rShutting down MQTT gracefully.')
 			# there can be cases where the connection is already gone.
 			try:
 				publish_result = self.mqttclient.publish(
@@ -232,7 +231,7 @@ class solmate_mqtt():
 				raise KeyboardInterrupt
 			if self.signal_reason == 2:
 				# the program was politely asked to terminate, we log and grant that request.
-				utils.logging('\rMQTT terminated on external request.', self.merged_config)
+				utils.logging('\rMQTT terminated on external request.')
 				sys.exit()
 
 	def _on_connect(self, client, userdata, flags, reason_code, properties = None):
@@ -252,7 +251,7 @@ class solmate_mqtt():
 				retain = True
 			)
 			self.connect_ok = True
-			utils.logging('MQTT is connected and running.', self.merged_config)
+			utils.logging('MQTT is connected and running.')
 			# we should always subscribe via on_connect callback to be sure
 			# the subscriptions are persisted across reconnections like client.subscribe("$SYS/#")
 			# technically it is not necessary that the topic is already available.
@@ -265,7 +264,7 @@ class solmate_mqtt():
 				self.send_sensor_update_message(self.remember_info_response, 'info')
 		else:
 			self.connect_ok = False
-			utils.logging('MQTT connection refused: ' + str(reason_code), self.merged_config)
+			utils.logging('MQTT connection refused: ' + str(reason_code))
 			self.mqttclient.loop_stop()
 
 	def _on_disconnect(self, client, userdata, disconnect_flags, reason_code, properties = None):
@@ -282,8 +281,8 @@ class solmate_mqtt():
 			utils.logging('MQTT disconnected: '
 				+ str(reason_code.getName())
 				+ ', packet: '
-				+ str(PacketTypes.Names[reason_code.packetType]),
-				self.merged_config)
+				+ str(PacketTypes.Names[reason_code.packetType])
+				)
 
 	def _on_publish(self, client, userdata, message, reason_codes, properties = None):
 		print(f'MQTT messages published: {message}')
@@ -386,7 +385,7 @@ class solmate_mqtt():
 
 			# first test, check if in range
 			if self._not_in_range(min_v, x, max_v):
-				utils.logging("Key " + "'" + str(key) + "': " + str(x) + " value out of range, using last valid: " + str(former), self.merged_config)
+				utils.logging("Key " + "'" + str(key) + "': " + str(x) + " value out of range, using last valid: " + str(former))
 				return former
 
 			# only if we have min/max injection
@@ -400,14 +399,14 @@ class solmate_mqtt():
 				if 'user_minimum_injection' in topic:
 					if x > max_e:
 						#print('min', x, max_e)
-						utils.logging("Key " + "'" + str(key) + "': " + str(x) + " value bigger than: " + str(max_e) + " using last valid: " + str(former), self.merged_config)
+						utils.logging("Key " + "'" + str(key) + "': " + str(x) + " value bigger than: " + str(max_e) + " using last valid: " + str(former))
 						return former
 
 				# next test, check if x < min_existing, it must be higher
 				if 'user_maximum_injection' in topic:
 					if x < min_e:
 						#print('max', x, min_v)
-						utils.logging("Key " + "'" + str(key) + "': " + str(x) + " value lower than: " + str(min_e) + " using last valid: " + str(former), self.merged_config)
+						utils.logging("Key " + "'" + str(key) + "': " + str(x) + " value lower than: " + str(min_e) + " using last valid: " + str(former))
 						return former
 			#print(x)
 			return x
@@ -417,7 +416,7 @@ class solmate_mqtt():
 			# which is definitely in the remember dictionary
 			# logging the issue with the fake name as this one is used in
 			# HA + the error cause + the errored value + the former valid value used
-			utils.logging("Key " + "'" + str(key) + "': " + str(err) + " using last valid: " + str(former), self.merged_config)
+			utils.logging("Key " + "'" + str(key) + "': " + str(err) + " using last valid: " + str(former))
 			return former
 
 	def _not_in_range(self, min_v, test, max_v):
@@ -438,7 +437,7 @@ class solmate_mqtt():
 			self.send_sensor_update_message(self.remember_info_response, 'info')
 			# new queue element = (route, key, value)
 			utils.mqtt_queue.put(('shutdown', {'shut_reboot': 'reboot'}))
-			utils.logging('Initializing SolMate Reboot.', self.merged_config)
+			utils.logging('Initializing SolMate Reboot.')
 
 	def send_sensor_update_message(self, response, endpoint):
 		# after connecting and initializing, this is the only location where we expect an error
@@ -455,7 +454,7 @@ class solmate_mqtt():
 				# fake a connected_to into the response if not present
 				response['connected_to'] = 'local' if self.api_available['local'] else 'cloud'
 				# fake a esham sw version (this program) into the response, for sure not present
-				response['esham_version'] = self.merged_config['internal_esham_version']
+				response['esham_version'] = utils.merged_config['internal_esham_version']
 
 			if endpoint == 'get_boost':
 				# remember the last response
@@ -487,8 +486,8 @@ class solmate_mqtt():
 
 	def set_operating_state_normal(self):
 		# do the cleanup after successful rebooting
-		utils.logging('SolMate has Rebooted.', self.merged_config)
-		utils.logging('Back to normal operation.', self.merged_config)
+		utils.logging('SolMate has Rebooted.')
+		utils.logging('Back to normal operation.')
 		self._init_button_command_topic('reboot', '')
 		self.eet_reboot_in_progress = False
 		self.send_sensor_update_message(self.remember_info_response, 'info')